support TypeScript project generation
