+++ packages/aio-commerce-lib-app/source/commands/generate/actions/constants.ts
@@ -37,67 +37,3 @@ export const COMMERCE_ACTION_INPUTS = Object.fromEntries(
 export const CUSTOM_IMPORTS_PLACEHOLDER = "// {{CUSTOM_SCRIPTS_IMPORTS}}";
 export const CUSTOM_SCRIPTS_MAP_PLACEHOLDER = "// {{CUSTOM_SCRIPTS_MAP}}";
 export const CUSTOM_SCRIPTS_LOADER_PLACEHOLDER = "// {{CUSTOM_SCRIPTS_LOADER}}";
-
-/** File extensions that require bundling or TypeScript web-src templates. */
-export const TYPESCRIPT_CONFIG_EXTENSIONS = new Set([".ts", ".mts", ".cts"]);
-
-export const JSX_FILE_EXTENSION = ".jsx";
-export const TSX_FILE_EXTENSION = ".tsx";
-
-export const APP_TITLE_PLACEHOLDER = "APP_TITLE";
-export const WEB_SOURCE_ENTRYPOINT_FILE = "index.html";
-export const WEB_SOURCE_IMPORT_ALIAS = "#web/*";
-
-/** Runtime dependencies required by the generated web-src app. */
-export const WEB_SOURCE_DEPENDENCIES = [
-  {
-    name: "@adobe/aio-commerce-lib-admin-ui",
-    version: __LIB_ADMIN_UI_RANGE__,
-  },
-  { name: "react", version: __REACT_VERSION__ },
-  { name: "react-dom", version: __REACT_DOM_VERSION__ },
-  { name: "@react-spectrum/s2", version: __SPECTRUM_S2_VERSION__ },
-] as const;
-
-/** Development dependencies required by the generated web-src app. */
-export const WEB_SOURCE_DEV_DEPENDENCIES = [
-  { name: "@types/react", version: __REACT_TYPES_VERSION__ },
-  { name: "@types/react-dom", version: __REACT_DOM_TYPES_VERSION__ },
-] as const;
-
-/** Development dependencies required only by TypeScript web-src scaffolds. */
-export const WEB_SOURCE_TS_DEV_DEPENDENCIES = [
-  { name: "@tsconfig/bases", version: "latest" },
-  { name: "typescript", version: "latest" },
-] as const;
-
-/** tsconfig.json generated for TypeScript web-src scaffolds. */
-export const WEB_SOURCE_TSCONFIG_FILE = "tsconfig.json";
-export const WEB_SOURCE_TSCONFIG = {
-  compilerOptions: {
-    allowImportingTsExtensions: true,
-    allowJs: true,
-    checkJs: true,
-    jsx: "react-jsx",
-    jsxImportSource: "react",
-    module: "esnext",
-    moduleResolution: "bundler",
-    noEmit: true,
-    rewriteRelativeImportExtensions: false,
-  },
-  extends: ["@tsconfig/bases/recommended"],
-
-  include: ["src/**/*.js", "src/**/*.ts", "src/**/*.tsx", "src/**/*.jsx"],
-};
-
-/** Parcel shared bundle config required by Spectrum S2 styles. */
-export const WEB_SOURCE_SHARED_BUNDLES = [
-  {
-    assets: [
-      "**/@react-spectrum/s2/**",
-      "src/commerce-backend-ui-2/web-src/*.{js,jsx,ts,tsx}",
-    ],
-    name: "s2-styles",
-    types: ["css"],
-  },
-];
