+++ packages/aio-commerce-lib-app/source/commands/generate/actions/main.ts
@@ -1,3 +1,15 @@
+/*
+ * Copyright 2026 Adobe. All rights reserved.
+ * This file is licensed to you under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License. You may obtain a copy
+ * of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under
+ * the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR REPRESENTATIONS
+ * OF ANY KIND, either express or implied. See the License for the specific language
+ * governing permissions and limitations under the License.
+ */
+
 import { CommerceSdkValidationError } from "@adobe/aio-commerce-lib-core/error";
 import { consola } from "consola";
 
@@ -6,6 +18,10 @@ import {
   CONFIGURATION_EXTENSION_POINT_ID,
   EXTENSIBILITY_EXTENSION_POINT_ID,
 } from "#commands/constants";
+import {
+  generateWebSrc,
+  prepareWebSourceImportAlias,
+} from "#commands/generate/web-src";
 import { loadAppManifest } from "#commands/utils";
 import { hasAdminUi, hasBusinessConfigSchema } from "#config/index";
 
@@ -13,9 +29,7 @@ import { getRuntimeActions } from "./config";
 import { TEMPLATES_DIR } from "./constants";
 import {
   generateActionFiles,
-  generateWebSrc,
   prepareRuntimeAppConfigModule,
-  prepareWebSourceImportAlias,
   updateExtConfig,
 } from "./lib";
 
@@ -68,7 +82,6 @@ export async function run(
       await prepareWebSourceImportAlias(extConfig);
       await generateWebSrc(
         extConfig,
-        BACKEND_UI_V2_EXTENSION_POINT_ID,
         appManifest.metadata.displayName,
         templatesDir,
       );
